[Silver II] Title: 과자 나눠주기, Time: 640 ms, Memory: 115180 KB -BaekjoonHub

+++ 백준/Silver/16401. 과자 나눠주기/과자 나눠주기.java
@@ -25,7 +25,7 @@ public class Main {
             int mid = (start+end) / 2;
 
             // 나눠줄 수 있는 갯수 구하기
-            long num = 0;
+            int num = 0;
             for(int i = N-1; i >= 0; i--){
                 if(snacks[i] < mid) continue;
                 num += snacks[i] / mid;
